pyproject.toml
[project]
name = "social-vision-agents"
version = "1.0.0"
description = "Production-ready social media content automation crew powered by LangGraph + Llama 4"
requires-python = ">=3.12"
authors = [
    { name = "HASHIRAMA21", email = "sodiaque806@gmail.com" },
]
dependencies = [
    # Orchestration
    "langgraph>=0.2.60",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain>=0.3.12",
    "langchain-core>=0.3.25",
    "langchain-community>=0.3.12",
    "langchain-ollama>=0.3.0",
    "langchain-groq>=0.2.4",
    "langchain-google-genai>=2.0",
    # Data & Config
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0",
    # RAG
    "chromadb>=0.5.23",
    # API
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    # Scheduler
    "apscheduler>=3.10.4",
    # Logging & Observability
    "structlog>=24.4",
    # HTTP client (for Grok Imagine API)
    "httpx>=0.28",
    # Async postgres
    "psycopg[binary]>=3.2",
    # Jinja2 for prompt templates
    "jinja2>=3.1",
    # Async PostgreSQL
    "asyncpg>=0.30",
    # Web search (Tavily primary, DuckDuckGo fallback)
    "tavily-python>=0.5",
    "duckduckgo-search>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
social-vision = "run_weekly:main_sync"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true